+++ connecteur-type/TdtConnecteur.class.php
@@ -77,9 +77,6 @@ abstract class TdtConnecteur extends Connecteur
 
     abstract public function verifClassif();
 
-    /* @deprecated PA 3.0 use sendHelios() instead */
-    abstract public function postHelios(DonneesFormulaire $donneesFormulaire);
-
     abstract public function sendHelios(Fichier $fichierHelios);
 
     abstract public function postActes(DonneesFormulaire $donneesFormulaire);