+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Suppression du modèle `Document`, remplacé par `DocumentSQL`
 - Suppression du modèle `Utilisateur`, remplacé par `UtilisateurSQL`
 - Suppression des fonctions SAEConnecteur::getAcuseReception et SAEConnecteur::getReply 
+- Suppression de TdtConnecteur::postHelios(), remplacé par TdtConnecteur::sendHelios()
 
 # [3.1.13] - 2022-04-18